+++ tests/validate/test_scenario_test.py
@@ -1,6 +1,7 @@
 import pytest
 
 from nucleus.validate import CreateScenarioTestError
+from nucleus.validate.constants import EntityLevel
 from nucleus.validate.scenario_test import ScenarioTest
 from tests.helpers import (
     EVAL_FUNCTION_COMPARISON,
@@ -48,7 +49,11 @@ def test_list_scenario_test(CLIENT, test_slice, annotations):
     CLIENT.validate.delete_scenario_test(scenario_test.id)
 
 
-def test_scenario_test_items(CLIENT, test_slice, slice_items, annotations):
+def test_scenario_test_get_dataset_items(
+    CLIENT,
+    test_slice,
+    slice_items,
+):
     test_name = "scenario_test_" + get_uuid()  # use uuid to make unique
 
     scenario_test = CLIENT.validate.create_scenario_test(
@@ -61,7 +66,31 @@ def test_scenario_test_items(CLIENT, test_slice, slice_items, annotations):
     actual_items_locations = [
         item.image_location for item in scenario_test.get_items()
     ]
-    assert expected_items_locations == actual_items_locations
+    assert set(actual_items_locations).issubset(expected_items_locations)
+    CLIENT.validate.delete_scenario_test(scenario_test.id)
+
+
+def test_scenario_test_get_scenes(
+    CLIENT, test_scene_slice, slice_scenes, annotations
+):
+    test_name = "scenario_test_" + get_uuid()  # use uuid to make unique
+
+    scenario_test = CLIENT.validate.create_scenario_test(
+        name=test_name,
+        slice_id=test_scene_slice.id,
+        evaluation_functions=[CLIENT.validate.eval_functions.bbox_iou()],
+    )
+
+    expected_scene_reference_ids = [
+        scene.reference_id for scene in slice_scenes
+    ]
+    actual_scene_reference_ids = [
+        scene.reference_id
+        for scene in scenario_test.get_items(level=EntityLevel.SCENE)
+    ]
+    assert set(actual_scene_reference_ids).issubset(
+        expected_scene_reference_ids
+    )
     CLIENT.validate.delete_scenario_test(scenario_test.id)